cache fix, metrix and cost-functions added

+++ lab1.1/src/main/java/lab1/ImageUtils.java
@@ -6,12 +6,12 @@ import org.opencv.imgcodecs.Imgcodecs;
 
 public class ImageUtils {
     @SuppressWarnings("all")
-    public static byte[][][] readImg(String path) {
+    public static int[][][] readImg(String path) {
 
         Imgcodecs imageCodecs = new Imgcodecs();
         Mat matrix = imageCodecs.imread(path);
 
-        byte[][][] image = new byte[matrix.rows()][matrix.cols()][3];
+        int[][][] image = new int[matrix.rows()][matrix.cols()][3];
         for (int i = 0; i < matrix.rows(); i++) {
             for (int j = 0; j < matrix.cols(); j++) {
                 image[i][j][0] = (byte) matrix.get(i,j)[0];
@@ -23,15 +23,15 @@ public class ImageUtils {
         return image;
     }
 
-    public static void writeImg(byte[][][] image, String path) {
+    public static void writeImg(int[][][] image, String path) {
         Mat matrix = new Mat(image.length,image[0].length,16);
 
         for (int i = 0; i < image.length; i++) {
             for (int j = 0; j < image[0].length; j++) {
                 byte[] pixel = new byte[3];
-                pixel[0] = image[i][j][0];
+                pixel[0] = (byte) image[i][j][0];
 //                pixel[1] = image[i][j];
-                pixel[2] = image[i][j][1];
+                pixel[2] = (byte) image[i][j][1];
                 matrix.put(i, j,pixel);
             }
         }
