+++ lab1.1/src/main/java/lab1/metrics/Difference.java
@@ -1,5 +1,5 @@
 package lab1.metrics;
 
 public interface Difference {
-     public double diff(byte[] p1, byte[] p2);
+     public double diff(int[] p1, int[] p2);
 }
